bpf: rename UINT8_MAX to UINT16_max and fix cluster_id casts
The UINT8_MAX macro is defined as the 16bit hex value 0xFFFF. Since
cluster_id is now a uint16 to allow for extended Cluster Mesh, this
renames the macro to UINT16_MAX, rather than changing the value.

Lingering casts of cluster_id as a uint8 have also been corrected to
ensure its usage as uint16.

Fixes: e8752ec125fc ("bpf: Define UINT8_MAX")

+++ bpf/bpf_lxc.c
@@ -1189,12 +1189,12 @@ skip_vtep:
 	if (!skip_tunnel) {
 		struct tunnel_key key = {};
 
-		if (cluster_id > UINT8_MAX)
+		if (cluster_id > UINT16_MAX)
 			return DROP_INVALID_CLUSTER_ID;
 
 		key.ip4 = ip4->daddr & IPV4_MASK;
 		key.family = ENDPOINT_KEY_IPV4;
-		key.cluster_id = (__u8)cluster_id;
+		key.cluster_id = (__u16)cluster_id;
 
 #if !defined(ENABLE_NODEPORT) && defined(ENABLE_HOST_FIREWALL)
 		/*
